+++ csrc/registers_api.c
@@ -112,15 +112,19 @@ int ul_reg_write(lua_State *L)
     case MSR_NONE:
         *((int_least64_t *)buffer) = value;
         break;
+#ifdef UC_X86_REG_MSR
     case MSR_X86:
         ((uc_x86_msr *)buffer)->rid = luaL_checkinteger(L, 3);
         ((uc_x86_msr *)buffer)->value = value;
         break;
+#endif
+#if defined(UC_ARM_REG_CP_REG) || defined(UC_ARM64_REG_CP_REG)
     case MSR_ARM:
     case MSR_ARM64:
         ((uc_arm_cp_reg *)buffer)->crn = luaL_checkinteger(L, 3);
         ((uc_arm_cp_reg *)buffer)->val = value;
         break;
+#endif
     default:
         UL_UNREACHABLE_MARKER;
     }
@@ -171,13 +175,17 @@ int ul_reg_read(lua_State *L)
 
     switch (msr_action)
     {
+#ifdef UC_X86_REG_MSR
     case MSR_X86:
         ((uc_x86_msr *)value_buffer)->rid = luaL_checkinteger(L, 3);
         break;
+#endif
+#if defined(UC_ARM_REG_CP_REG) || defined(UC_ARM64_REG_CP_REG)
     case MSR_ARM:
     case MSR_ARM64:
         ((uc_arm_cp_reg *)value_buffer)->crn = luaL_checkinteger(L, 3);
         break;
+#endif
     case MSR_NONE:
         break;
     default: