+++ CHANGELOG.rst
@@ -14,6 +14,12 @@ Bugs
   register ID was the first member of the struct used for reading/writing.
 * Crash deterministically when attempting to access an X86 MSR with
   ``reg_read_batch()``, ``reg_read_batch_as()``, and ``reg_write_batch()``.
+* Fix wrong hook function being called for CPUID. It wasn't being special-cased,
+  meaning the normal instruction hook function was called. This returns void
+  when Unicorn was expecting a boolean, so whether Unicorn fell back to using
+  the default behavior was undefined.
+* Crash with a better error message when attempting to hook CPUID on a
+  version of Unicorn that doesn't support it.
 
 .. _issue 47: https://github.com/dargueta/unicorn-lua/issues/47
 .. _issue 48: https://github.com/dargueta/unicorn-lua/issues/48